Specify message to one channel
