+++ src/types/SendResult.ts
@@ -1,0 +1,7 @@
+import type { Channel } from './Channel';
+
+export interface SendResult {
+  channel: Channel;
+  success: boolean;
+  recipientCount: number;
+}
